Marquis updates and moving code around
Changed the way the marquis works - not entirely finished, but now it is
always scrolling. 5 second freezes at the beginning and end of the
filename to promote readability.

Put some stuff in different classes so it wasn't clogging up Form1.cs,
which has been renamed.

Fixed some bugs involving the scroll timer implementation, the next/back
buttons, and the filename of the currently displayed image.